3265: Deletion of StateStoreArrowFileStore v1 and replace with v2
