+++ java/statestorev2/src/main/java/sleeper/statestorev2/transactionlog/snapshots/DynamoDBTransactionLogSnapshotCreator.java
@@ -32,7 +32,7 @@ import sleeper.core.statestore.transactionlog.transaction.FileReferenceTransacti
 import sleeper.core.statestore.transactionlog.transaction.PartitionTransaction;
 import sleeper.core.statestore.transactionlog.transaction.TransactionSerDeProvider;
 import sleeper.core.table.TableStatus;
-import sleeper.statestorev2.StateStoreArrowFileStoreV2;
+import sleeper.statestorev2.StateStoreArrowFileStore;
 import sleeper.statestorev2.transactionlog.DuplicateSnapshotException;
 import sleeper.statestorev2.transactionlog.DynamoDBTransactionLogStore;
 import sleeper.statestorev2.transactionlog.S3TransactionBodyStore;
@@ -52,7 +52,7 @@ public class DynamoDBTransactionLogSnapshotCreator {
     private final TransactionBodyStore transactionBodyStore;
     private final LatestSnapshotsMetadataLoader latestMetadataLoader;
     private final DynamoDBTransactionLogSnapshotSaver snapshotSaver;
-    private final StateStoreArrowFileStoreV2 fileStore;
+    private final StateStoreArrowFileStore fileStore;
 
     /**
      * Builds a snapshot creator for a given Sleeper table.
@@ -91,7 +91,7 @@ public class DynamoDBTransactionLogSnapshotCreator {
         this.latestMetadataLoader = latestMetadataLoader;
         this.snapshotSaver = new DynamoDBTransactionLogSnapshotSaver(
                 latestMetadataLoader, metadataSaver, instanceProperties, tableProperties, configuration, s3Client, s3TransferManager);
-        this.fileStore = new StateStoreArrowFileStoreV2(instanceProperties, tableProperties, s3Client, s3TransferManager);
+        this.fileStore = new StateStoreArrowFileStore(instanceProperties, tableProperties, s3Client, s3TransferManager);
     }
 
     /**
